refactor: initial split in bin+lib
